requirements.txt
slackclient==2.9.4
functions-framework==3.2.0
google-cloud-logging==3.5.0
google_cloud_bigquery==3.11.4
Werkzeug==2.3.7
